+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - add support for sdwan_system_ipv4_device_access_feature resource
 - add support for sdwan_system_ipv6_device_access_feature resource
 - add support for sdwan_service_dhcp_server_feature resource
+- add support for sdwan_configuration_group resource
+- add support for sdwan_tag
 
 ## 1.0.0
 
@@ -49,7 +51,6 @@
 - add support for sdwan_cellular_cedge_profile_feature_template resource
 - add support for sdwan_cisco_vpn_interface_gre_feature_template resource
 - add support for sdwan_vpn_interface_cellular_feature_template resource
-- add support for sdwan_configuration_group resource (single device configuration group only)
 
 ## 0.1.0